Cast and create real numbers based on valid float precision

+++ include/hpp/fcl/data_types.h
@@ -41,28 +41,30 @@
 #include <Eigen/Core>
 #include <Eigen/Geometry>
 
+#include <cstdint>
+
 #include <hpp/fcl/config.hh>
 
 namespace hpp {
 
 #ifdef HPP_FCL_HAS_OCTOMAP
-#define OCTOMAP_VERSION_AT_LEAST(x, y, z) \
-  (OCTOMAP_MAJOR_VERSION > x ||           \
-   (OCTOMAP_MAJOR_VERSION >= x &&         \
-    (OCTOMAP_MINOR_VERSION > y ||         \
+#define OCTOMAP_VERSION_AT_LEAST(x, y, z)                                      \
+  (OCTOMAP_MAJOR_VERSION > x ||                                                \
+   (OCTOMAP_MAJOR_VERSION >= x &&                                              \
+    (OCTOMAP_MINOR_VERSION > y ||                                              \
      (OCTOMAP_MINOR_VERSION >= y && OCTOMAP_PATCH_VERSION >= z))))
 
-#define OCTOMAP_VERSION_AT_MOST(x, y, z) \
-  (OCTOMAP_MAJOR_VERSION < x ||          \
-   (OCTOMAP_MAJOR_VERSION <= x &&        \
-    (OCTOMAP_MINOR_VERSION < y ||        \
+#define OCTOMAP_VERSION_AT_MOST(x, y, z)                                       \
+  (OCTOMAP_MAJOR_VERSION < x ||                                                \
+   (OCTOMAP_MAJOR_VERSION <= x &&                                              \
+    (OCTOMAP_MINOR_VERSION < y ||                                              \
      (OCTOMAP_MINOR_VERSION <= y && OCTOMAP_PATCH_VERSION <= z))))
-#endif  // HPP_FCL_HAS_OCTOMAP
-}  // namespace hpp
+#endif // HPP_FCL_HAS_OCTOMAP
+} // namespace hpp
 
 namespace hpp {
 namespace fcl {
-typedef double FCL_REAL;
+typedef float FCL_REAL;
 typedef Eigen::Matrix<FCL_REAL, 3, 1> Vec3f;
 typedef Eigen::Matrix<FCL_REAL, Eigen::Dynamic, 1> VecXf;
 typedef Eigen::Matrix<FCL_REAL, 3, 3> Matrix3f;
@@ -94,7 +96,7 @@ enum GJKConvergenceCriterionType { Relative, Absolute };
 
 /// @brief Triangle with 3 indices for points
 class HPP_FCL_DLLAPI Triangle {
- public:
+public:
   typedef std::size_t index_type;
   typedef int size_type;
 
@@ -114,18 +116,18 @@ class HPP_FCL_DLLAPI Triangle {
   /// @brief Access the triangle index
   inline index_type operator[](index_type i) const { return vids[i]; }
 
-  inline index_type& operator[](index_type i) { return vids[i]; }
+  inline index_type &operator[](index_type i) { return vids[i]; }
 
   static inline size_type size() { return 3; }
 
-  bool operator==(const Triangle& other) const {
+  bool operator==(const Triangle &other) const {
     return vids[0] == other.vids[0] && vids[1] == other.vids[1] &&
            vids[2] == other.vids[2];
   }
 
-  bool operator!=(const Triangle& other) const { return !(*this == other); }
+  bool operator!=(const Triangle &other) const { return !(*this == other); }
 
- private:
+private:
   /// @brief indices for each vertex of triangle
   index_type vids[3];
 };
@@ -152,25 +154,25 @@ struct HPP_FCL_DLLAPI Quadrilateral {
   /// @access the quadrilateral index
   inline index_type operator[](index_type i) const { return vids[i]; }
 
-  inline index_type& operator[](index_type i) { return vids[i]; }
+  inline index_type &operator[](index_type i) { return vids[i]; }
 
   static inline size_type size() { return 4; }
 
-  bool operator==(const Quadrilateral& other) const {
+  bool operator==(const Quadrilateral &other) const {
     return vids[0] == other.vids[0] && vids[1] == other.vids[1] &&
            vids[2] == other.vids[2] && vids[3] == other.vids[3];
   }
 
-  bool operator!=(const Quadrilateral& other) const {
+  bool operator!=(const Quadrilateral &other) const {
     return !(*this == other);
   }
 
- private:
+private:
   index_type vids[4];
 };
 
-}  // namespace fcl
+} // namespace fcl
 
-}  // namespace hpp
+} // namespace hpp
 
 #endif
